Add test for `editTask`
